Split parser into separate file, add tests
